--- App/Navigation/CustomNavBar.js
@@ -1,103 +1,0 @@
-import React, { PropTypes } from 'react';
-import { View, Image, LayoutAnimation } from 'react-native';
-import NavItems from './NavItems';
-import styles from './Styles/CustomNavBarStyle';
-import SearchBar from '../Components/SearchBar';
-import { connect } from 'react-redux';
-import { metrics, Images } from '../Themes';
-import SearchActions from '../Redux/SearchRedux';
-
-class CustomNavBar extends React.Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      showSearchBar: false,
-    };
-  }
-
-  showSearchBar = () => {
-    this.setState({showSearchBar: true});
-  }
-
-  cancelSearch = () => {
-    this.setState({showSearchBar: false});
-    this.props.cancelSearch();
-  }
-
-  onSearch = (searchTerm) => {
-    this.props.performSearch(searchTerm);
-  }
-
-  renderMiddle() {
-    LayoutAnimation.configureNext(LayoutAnimation.Presets.easeInEaseOut);
-    if (this.state.showSearchBar) {
-      return <SearchBar onSearch={this.props.performSearch} searchTerm={this.props.searchTerm} onCancel={this.cancelSearch} />;
-    }
-    return (
-      <Image resizeMode="cover" style={styles.logo} source={Images.clearLogo} />
-    );
-  }
-
-  renderRightButtons() {
-    if (this.state.showSearchBar) {
-      return <View style={{width: metrics.icons.medium}} />;
-    }
-    return (
-      <View style={styles.rightButtons}>
-        {NavItems.searchButton(this.showSearchBar)}
-      </View>
-    );
-  }
-
-  renderLeftButtons() {
-    if (this.state.showSearchBar) {
-      return null;
-    }
-    return (
-      <View style={styles.leftButtons}>
-        {NavItems.backButton()}
-      </View>
-    );
-  }
-
-  render() {
-    let state = this.props.navigationState;
-    let selected = state.children[state.index];
-    while (selected.hasOwnProperty('children')) {
-      state = selected;
-      selected = selected.children[selected.index];
-    }
-
-    const containerStyle = [
-      styles.container,
-      this.props.navigationBarStyle,
-      state.navigationBarStyle,
-      selected.navigationBarStyle,
-    ];
-
-    return (
-      <View style={containerStyle}>
-        {this.renderLeftButtons()}
-        {this.renderMiddle()}
-        {this.renderRightButtons()}
-      </View>
-    );
-  }
-}
-
-CustomNavBar.propTypes = {
-  navigationState: PropTypes.object,
-  navigationBarStyle: View.propTypes.style,
-};
-
-const mapStateToProps = state => ({
-  searchTerm: state.search.searchTerm,
-});
-
-const mapDispatchToProps = dispatch => ({
-  performSearch: searchTerm => dispatch(SearchActions.search(searchTerm)),
-  cancelSearch: () => dispatch(SearchActions.cancelSearch()),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(CustomNavBar);
