--- App/Services/styles.js
@@ -1,5 +1,0 @@
-// @flow
-import { Dimensions } from 'react-native';
-
-export const viewportWidth = (percentageWidth: number) => Dimensions.get('window').width * (percentageWidth / 100);
-export const viewportHeight = (percentageHeight: number) => Dimensions.get('window').height * (percentageHeight / 100);
